requirements.txt
streamlit>=1.35,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
openai>=1.40,<2.0
pydantic>=2.6,<3.0
